requirements.txt
Flask~=2.3.2
flask_socketio
eventlet
python-socketio[client]
